% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaible_syllable_funs.R
\name{avaible_syllable_funs}
\alias{avaible_syllable_funs}
\title{Available Syllable Functions}
\usage{
avaible_syllable_funs(ncols = 4, ...)
}
\arguments{
\item{ncols}{The number of columns to use.}

\item{\ldots}{Other arguments passed to \code{\link[base]{matrix}}.}
}
\value{
Returns a \code{\link[base]{character}} vector,
\code{\link[base]{matrix}} of all syllable functions, or a
\code{\link[base]{list}} of syllable functions by type.
}
\description{
See a listing of all available syllable functions.
}
\examples{
avaible_syllable_funs()

avaible_syllable_funs(ncols=3)
avaible_syllable_funs(1)
avaible_syllable_funs(byrow = TRUE)
}
\keyword{type}

