% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable-package.R
\docType{data}
\name{common_polysyllabic_proper_nouns}
\alias{common_polysyllabic_proper_nouns}
\title{Common Proper Nouns}
\format{A character vector with 250 elements}
\usage{
data(common_polysyllabic_proper_nouns)
}
\description{
A dataset containing a character vector of common proper nouns Greater than 2
syllables.  This is useful to exclude words for "complex word" detection
used in the Gunning fog index.
}
\keyword{datasets}

