% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_vector.R
\name{count_vector}
\alias{count_vector}
\title{Vector Syllable Counts}
\usage{
count_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a list of vectors of integer counts.
}
\description{
Syllable counts for the words in a vector of strings.
}
\examples{
sents <- c("I like chicken.", "I want eggs benidict for breakfast.")
count_vector(sents)

Map(function(x, y) setNames(x, y),
   count_vector(sents),
   strsplit(gsub("[^a-z ]", "", tolower(sents)), "\\\\s+")
)

count_vector(presidential_debates_2012$dialogue)
}

