% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_vector_by.R
\name{count_vector_by}
\alias{count_vector_by}
\title{Vector Syllable Counts By Grouping Variable(s)}
\usage{
count_vector_by(x, group, ...)
}
\arguments{
\item{x}{A character vector.}

\item{group}{The grouping variable(s).  Takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{\ldots}{ignored}
}
\value{
Returns a named list of vectors of integer counts for each grouping variable.
}
\description{
Syllable counts for the words in a vector of strings by grouping variables.
}
\examples{
dat <- data.frame(
   text = c("I like chicken.", "I want eggs benidict for breakfast.", "Really?"),
   group = c("A", "B", "A")
)
count_vector_by(dat$text, dat$group)

with(presidential_debates_2012, count_vector_by(dialogue, list(person, time)))
}

