% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable-package.R
\docType{data}
\name{hamlets_soliloquy}
\alias{hamlets_soliloquy}
\title{Hamlet's Soliloquy}
\format{A character vector with 1 element}
\usage{
data(hamlets_soliloquy)
}
\description{
A dataset containing a character string of Hamlet's soliloquy.
}
\keyword{datasets}

