% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avaible_syllable_funs.R
\name{print.available}
\alias{print.available}
\title{Prints an available object.}
\usage{
\method{print}{available}(x, ...)
}
\arguments{
\item{x}{The available object}

\item{\ldots}{ignored}
}
\description{
Prints an available object.
}

