% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readability_word_stats.R
\name{readability_word_stats}
\alias{readability_word_stats}
\title{Readability Word Statistics}
\usage{
readability_word_stats(x, as.tibble = tibble_output(), ...)
}
\arguments{
\item{x}{A character vector.}

\item{as.tibble}{logical.  If \code{TRUE} the output class will be set to a
\pkg{tibble}, otherwise a \code{\link[data.table]{data.table}}.  Default
checks \code{getOption("tibble.out")} as a logical.  If this is \code{NULL}
the default \code{\link[textshape]{tibble_output}} will set \code{as.tibble}
to \code{TRUE} if \pkg{dplyr} is loaded.  Otherwise, the output will be a
\code{\link[data.table]{data.table}}.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{\link[base]{data.frame}}
(\code{\link[data.table]{data.table}}) readability word statistics.
}
\description{
Word statistics commonly used to calculate readability sores.
}
\examples{
x <- c("I like excellent chicken.", "I want eggs Benedict now.", "Really?",
    "I thought that was good.  Well a bit weird.", "I know right!")
readability_word_stats(x)

\dontrun{
readability_word_stats(presidential_debates_2012$dialogue)
}
}

