% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textshape_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{set_output}
\alias{tibble_output}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{textshape}{\code{\link[textshape]{tibble_output}}, \code{\link[textshape]{set_output}}}
}}
\keyword{internal}

