% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_string.R
\name{sum_string}
\alias{sum_string}
\title{String Syllables Sum}
\usage{
sum_string(x, ...)
}
\arguments{
\item{x}{A character string.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of syllables in the string.
}
\description{
Syllable sum for the words in a single string.
}
\examples{
sum_string("I like chicken and eggs for breakfast")
sum_string(hamlets_soliloquy)
}
\seealso{
Other syllable.functions: \code{\link{count_string}},
  \code{\link{count_vector_by}},
  \code{\link{count_vector}}, \code{\link{sum_vector_by}},
  \code{\link{sum_vector}},
  \code{\link{tally_both_string}},
  \code{\link{tally_both_vector_by}},
  \code{\link{tally_both_vector}},
  \code{\link{tally_di_string}},
  \code{\link{tally_di_vector_by}},
  \code{\link{tally_di_vector}},
  \code{\link{tally_mono_string}},
  \code{\link{tally_mono_vector_by}},
  \code{\link{tally_mono_vector}},
  \code{\link{tally_poly_string}},
  \code{\link{tally_poly_vector_by}},
  \code{\link{tally_poly_vector}},
  \code{\link{tally_short_string}},
  \code{\link{tally_short_vector_by}},
  \code{\link{tally_short_vector}}
}

