% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sum_vector.R
\name{sum_vector}
\alias{sum_vector}
\title{Vector Syllable Sums}
\usage{
sum_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a list of vectors of integer counts.
}
\description{
Syllable sums for the words in a vector of strings.
}
\examples{
sents <- c("I like chicken.", "I want eggs benidict for breakfast.")
sum_vector(sents)

data.frame(
   sents, syllable_count = sum_vector(sents),
   stringsAsFactors = FALSE
)

sum_vector(presidential_debates_2012$dialogue)
}

