% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllable-package.R
\docType{package}
\name{syllable}
\alias{package-syllable}
\alias{syllable}
\alias{syllable-package}
\title{A Small Collection of Syllable Counting Functions}
\description{
Tools for counting syllables and polysyllables.  The tools rely primarily on
a \pkg{data.table} hash table lookup, resulting in fast syllable counting.
}

