% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syllable-package.R
\docType{data}
\name{syllable_counts_data}
\alias{syllable_counts_data}
\title{A \pkg{data.table} of Words and Syllable Counts}
\format{A data frame with 124603 rows and 2 variables}
\usage{
data(syllable_counts_data)
}
\description{
A \pkg{data.table} hash table dataset containing words and syllable counts.
}
\details{
\itemize{
  \item word. A character coolumn of lower case words.
  \item syllables. The syllable counts per word.
}
}
\references{
Counts scraped from \url{http://www.poetrysoup.com}
}
\keyword{datasets}

