% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_both_string.R
\name{tally_both_string}
\alias{tally_both_string}
\title{String Syllable Tally of Non-Polysyllabic and Polysyllabic Words}
\usage{
tally_both_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a two integer vector of the total number of non-polysyllables
(non) and polysyllables (poly) in the string.
}
\description{
Non-polysyllabic and polysyllabic word tallies for the words in a string.
}
\examples{
tally_both_string("I like excellent chicken and mediocre hotdogs!")
tally_both_string(hamlets_soliloquy)
}

