% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_both_vector.R
\name{tally_both_vector}
\alias{tally_both_vector}
\title{Vector Tally of Short-Syllabic and Poly-Syllabic Words}
\usage{
tally_both_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a two column \code{\link[base]{data.frame}}
(\code{\link[data.table]{data.table}}) of integer tallies for the total number of
short syllable (short) and poly syllable (poly) words for each string in the vector.
}
\description{
Short-syllabic (< 3 syllables) and poy-syllabic word tallies for the words in
a vector of strings.
}
\examples{
sents <- c("I like excellent chicken.", "I want eggs benedict for Festivus.")
tally_both_vector(sents)
tally_both_vector(presidential_debates_2012$dialogue)
}

