% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_di_string.R
\name{tally_di_string}
\alias{tally_di_string}
\title{String Syllable Tally of Disyllabic Words}
\usage{
tally_di_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of disyllables in the
string.
}
\description{
Disyllabic word tallies for the words in a string.
}
\examples{
tally_di_string("I like chicken and hotdogs!")
tally_di_string(hamlets_soliloquy)
}

