% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_di_vector.R
\name{tally_di_vector}
\alias{tally_di_vector}
\title{Vector Tally of Di-syllabic Words}
\usage{
tally_di_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a vector of integer tallies for the total number of
disyllable words for each string in the vector.
}
\description{
Di-syllabic word tallies for the words in a vector of strings.
}
\examples{
sents <- c("I like chicken.", "I want eggs benedict for breakfast.")
tally_di_vector(sents)
tally_di_vector(presidential_debates_2012$dialogue)
}

