% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_mono_string.R
\name{tally_mono_string}
\alias{tally_mono_string}
\title{String Syllable Tally of Monosyllabic Words}
\usage{
tally_mono_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of monosyllables in the string.
}
\description{
Monosyllabic word tallies for the words in a string.
}
\examples{
tally_mono_string("I like chicken and beans!")
tally_mono_string(hamlets_soliloquy)
}

