% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_mono_vector.R
\name{tally_mono_vector}
\alias{tally_mono_vector}
\title{Vector Tally of Mono-syllabic Words}
\usage{
tally_mono_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a vector of integer tallies for the total number of
monosyllable words for each string in the vector.
}
\description{
Mono-syllabic word tallies for the words in a vector of strings.
}
\examples{
sents <- c("I like chicken.", "I want eggs benedict for breakfast.")
tally_mono_vector(sents)
tally_mono_vector(presidential_debates_2012$dialogue)
}

