% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_mono_vector_by.R
\name{tally_mono_vector_by}
\alias{tally_mono_vector_by}
\title{Vector Tally of Mono-Syllabic Words By Grouping Variable(s)}
\usage{
tally_mono_vector_by(x, group, ...)
}
\arguments{
\item{x}{A character vector.}

\item{group}{The grouping variable(s).  Takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{\ldots}{ignored}
}
\value{
Returns a \code{\link[base]{data.frame}}
(\code{\link[data.table]{data.table}}) of integer tallies for the total number of
mono syllable words for each string in the vector.
}
\description{
Mono-syllabic word tallies for the words in a vector of strings by grouping
variable(s).
}
\examples{
dat <- data.frame(
   text = c("I like chicken.", "I want eggs benedict for breakfast.", "Really?"),
   group = c("A", "B", "A")
)
tally_mono_vector_by(dat$text, dat$group)

with(presidential_debates_2012, tally_mono_vector_by(dialogue, person))
with(presidential_debates_2012, tally_mono_vector_by(dialogue, list(role, time)))
with(presidential_debates_2012, tally_mono_vector_by(dialogue, list(person, time)))
}

