% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_poly_string.R
\name{tally_poly_string}
\alias{tally_poly_string}
\title{String Syllable Tally of Polyllabic Words}
\usage{
tally_poly_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of polysyllables in the
string.
}
\description{
Polyllabic word tallies for the words in a string.
}
\examples{
tally_poly_string("I like generalized chicken and tarnishing hotdogs!")
tally_poly_string(hamlets_soliloquy)
}

