% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_short_vector.R
\name{tally_short_vector}
\alias{tally_short_vector}
\title{Vector Tally of Short-syllabic and Poly-syllabic Words}
\usage{
tally_short_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a vector of integer tallies for the total number of
short syllable words for each string in the vector.
}
\description{
Short-syllabic (< 3 syllables) and poly-syllabic word tallies for the words
in a vector of strings.
}
\examples{
sents <- c("I like excellent chicken.", "I want eggs benedict for breakfast.")
tally_short_vector(sents)
tally_short_vector(presidential_debates_2012$dialogue)
}

