% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_short_vector_by.R
\name{tally_short_vector_by}
\alias{tally_short_vector_by}
\title{Vector Tally of Short-Syllabic Words By Grouping Variable(s)}
\usage{
tally_short_vector_by(x, group, ...)
}
\arguments{
\item{x}{A character vector.}

\item{group}{The grouping variable(s).  Takes a single grouping variable or a
list of 1 or more grouping variables.}

\item{\ldots}{ignored}
}
\value{
Returns a \code{\link[base]{data.frame}}
(\code{\link[data.table]{data.table}}) of integer tallies for the total number
of short syllable words for each string in the vector.
}
\description{
Short-syllabic (< 3 syllables) word tallies for the words in a vector of
strings by grouping variable(s).
}
\examples{
dat <- data.frame(
   text = c("I like excellent chicken.", "I want eggs benedict now.", "Really?"),
   group = c("A", "B", "A")
)
tally_short_vector_by(dat$text, dat$group)

with(presidential_debates_2012, tally_short_vector_by(dialogue, person))
with(presidential_debates_2012, tally_short_vector_by(dialogue, list(role, time)))
with(presidential_debates_2012, tally_short_vector_by(dialogue, list(person, time)))
}
\seealso{
Other syllable.functions: \code{\link{count_string}};
  \code{\link{count_vector_by}};
  \code{\link{count_vector}}; \code{\link{sum_string}};
  \code{\link{sum_vector_by}}; \code{\link{sum_vector}};
  \code{\link{tally_both_string}};
  \code{\link{tally_both_vector_by}};
  \code{\link{tally_both_vector}};
  \code{\link{tally_di_string}};
  \code{\link{tally_di_vector_by}};
  \code{\link{tally_di_vector}};
  \code{\link{tally_mono_string}};
  \code{\link{tally_mono_vector_by}};
  \code{\link{tally_mono_vector}};
  \code{\link{tally_poly_string}};
  \code{\link{tally_poly_vector_by}};
  \code{\link{tally_poly_vector}};
  \code{\link{tally_short_string}};
  \code{\link{tally_short_vector}}
}

