% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_string.R
\name{count_string}
\alias{count_string}
\title{String Syllable Counts}
\usage{
count_string(x, ...)
}
\arguments{
\item{x}{A character string.}

\item{\ldots}{ignored}
}
\value{
Returns a vector of integer counts.
}
\description{
Syllable counts for the words in a single string.
}
\examples{
count_string("I like chicken and eggs for breakfast")
count_string(hamlets_soliloquy)
library(stringi)

data.frame(
    word = stri_extract_all_words(stri_trans_tolower(hamlets_soliloquy))[[1]],
    syllables = count_string(hamlets_soliloquy)
)
}

