% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_vector.R
\name{count_vector}
\alias{count_vector}
\title{Vector Syllable Counts}
\usage{
count_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a list of vectors of integer counts.
}
\description{
Syllable counts for the words in a vector of strings.
}
\examples{
sents <- c("I like chicken.", "I want eggs benidict for breakfast.")
count_vector(sents)

Map(function(x, y) setNames(x, y),
   count_vector(sents),
   strsplit(gsub("[^a-z ]", "", tolower(sents)), "\\\\s+")
)

count_vector(presidential_debates_2012$dialogue)
}
\seealso{
Other syllable.functions: \code{\link{count_string}};
  \code{\link{count_vector_by}}; \code{\link{sum_string}};
  \code{\link{sum_vector_by}}; \code{\link{sum_vector}};
  \code{\link{tally_both_string}};
  \code{\link{tally_both_vector_by}};
  \code{\link{tally_both_vector}};
  \code{\link{tally_di_string}};
  \code{\link{tally_di_vector_by}};
  \code{\link{tally_di_vector}};
  \code{\link{tally_mono_string}};
  \code{\link{tally_mono_vector_by}};
  \code{\link{tally_mono_vector}};
  \code{\link{tally_poly_string}};
  \code{\link{tally_poly_vector_by}};
  \code{\link{tally_poly_vector}};
  \code{\link{tally_short_string}};
  \code{\link{tally_short_vector_by}};
  \code{\link{tally_short_vector}}
}

