% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sum_string.R
\name{sum_string}
\alias{sum_string}
\title{String Syllables Sum}
\usage{
sum_string(x, ...)
}
\arguments{
\item{x}{A character string.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of syllables in the string.
}
\description{
Syllable sum for the words in a single string.
}
\examples{
sum_string("I like chicken and eggs for breakfast")
sum_string(hamlets_soliloquy)
}

