% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sum_vector_by.R
\name{sum_vector_by}
\alias{sum_vector_by}
\title{Vector Syllable Sums By Grouping Variable(s)}
\usage{
sum_vector_by(x, group, ...)
}
\arguments{
\item{x}{A character vector.}

\item{group}{The grouping variable(s).  Takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{\ldots}{ignored}
}
\value{
Returns a \code{\link[base]{data.frame}} (\code{\link[data.table]{data.table}}) of
syllable sums by grouping variable.
}
\description{
Syllable sums for the words in the vector(s) of strings for grouping variables.
}
\examples{
dat <- data.frame(
   text = c("I like chicken.", "I want eggs benedict for breakfast.", "Really?"),
   group = c("A", "B", "A")
)
sum_vector_by(dat$text, dat$group)

with(presidential_debates_2012, sum_vector_by(dialogue, list(person, time)))
}

