% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_both_string.R
\name{tally_both_string}
\alias{tally_both_string}
\title{String Syllable Tally of Short-syllabic and Poly-syllabic Words}
\usage{
tally_both_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a two integer vector of the total number of non-polysyllables
(short) and polysyllables (poly) in the string.
}
\description{
Short-syllabic (< 3 syllables) and poly-syllabic word tallies for the words
in a string.
}
\examples{
tally_both_string("I like excellent chicken and mediocre hotdogs!")
tally_both_string(hamlets_soliloquy)
}

