% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_both_string.R
\name{tally_both_string}
\alias{tally_both_string}
\title{String Syllable Tally of Short-syllabic and Poly-syllabic Words}
\usage{
tally_both_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored.}
}
\value{
Returns a two integer vector of the total number of non-polysyllables
(short) and polysyllables (poly) in the string.
}
\description{
Short-syllabic (< 3 syllables) and poly-syllabic word tallies for the words
in a string.
}
\examples{
tally_both_string("I like excellent chicken and mediocre hotdogs!")
tally_both_string(hamlets_soliloquy)
}
\seealso{
Other syllable.functions: \code{\link{count_string}},
  \code{\link{count_vector_by}},
  \code{\link{count_vector}}, \code{\link{sum_string}},
  \code{\link{sum_vector_by}}, \code{\link{sum_vector}},
  \code{\link{tally_both_vector_by}},
  \code{\link{tally_both_vector}},
  \code{\link{tally_di_string}},
  \code{\link{tally_di_vector_by}},
  \code{\link{tally_di_vector}},
  \code{\link{tally_mono_string}},
  \code{\link{tally_mono_vector_by}},
  \code{\link{tally_mono_vector}},
  \code{\link{tally_poly_string}},
  \code{\link{tally_poly_vector_by}},
  \code{\link{tally_poly_vector}},
  \code{\link{tally_short_string}},
  \code{\link{tally_short_vector_by}},
  \code{\link{tally_short_vector}}
}

