% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_mono_string.R
\name{tally_mono_string}
\alias{tally_mono_string}
\title{String Syllable Tally of Monosyllabic Words}
\usage{
tally_mono_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of monosyllables in the string.
}
\description{
Monosyllabic word tallies for the words in a string.
}
\examples{
tally_mono_string("I like chicken and beans!")
tally_mono_string(hamlets_soliloquy)
}
\seealso{
Other syllable.functions: \code{\link{count_string}};
  \code{\link{count_vector_by}};
  \code{\link{count_vector}}; \code{\link{sum_string}};
  \code{\link{sum_vector_by}}; \code{\link{sum_vector}};
  \code{\link{tally_both_string}};
  \code{\link{tally_both_vector_by}};
  \code{\link{tally_both_vector}};
  \code{\link{tally_di_string}};
  \code{\link{tally_di_vector_by}};
  \code{\link{tally_di_vector}};
  \code{\link{tally_mono_vector_by}};
  \code{\link{tally_mono_vector}};
  \code{\link{tally_poly_string}};
  \code{\link{tally_poly_vector_by}};
  \code{\link{tally_poly_vector}};
  \code{\link{tally_short_string}};
  \code{\link{tally_short_vector_by}};
  \code{\link{tally_short_vector}}
}

