% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_poly_vector.R
\name{tally_poly_vector}
\alias{tally_poly_vector}
\title{Vector Tally of Poly-syllabic Words}
\usage{
tally_poly_vector(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a vector of integer tallies for the total number of
polysyllable words for each string in the vector.
}
\description{
Poly-syllabic word tallies for the words in a vector of strings.
}
\examples{
sents <- c("I like excellent chicken.", "I desparately want eggs benedict.")
tally_poly_vector(sents)
tally_poly_vector(presidential_debates_2012$dialogue)
}

