% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_short_string.R
\name{tally_short_string}
\alias{tally_short_string}
\title{String Syllable Tally of Non-Polysyllabic (< 3 syllables) Words}
\usage{
tally_short_string(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{\ldots}{ignored}
}
\value{
Returns a single integer of the total number of non-polysyllables in
the string.
}
\description{
Non-Polysyllabic word tallies for the words in a string.
}
\examples{
tally_short_string("I like excellent chicken and blah hotdogs!")
tally_short_string(hamlets_soliloquy)
}

